pyproject.toml
[project]
name = "lif-main"
version = "0.1.0"
description = "Monorepo for LIF Components"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "dagster~=1.11",
    "dagster-graphql~=1.11",
    "deepdiff~=8.6",
    "fastapi~=0.115",
    "fastmcp~=2.9",
    "httpx~=0.28",
    "jsonata-python~=0.6",
    "jsonpath-ng~=1.7",
    "jsonref~=1.1",
    "jsonschema~=4.25",
    "langchain~=0.3",
    "langchain-mcp-adapters>=0.1.14,<0.2.0",
    "langchain-openai~=0.3",
    "langgraph~=0.4",
    "langgraph-prebuilt~=0.2",
    "langmem~=0.0",
    "mcp>=1.10,<2.0",
    "mcp-graphql~=0.3",
    "mkdocs-material~=9.6",
    "numpy~=2.3",
    "pandas~=2.2",
    "pydantic~=2.11",
    "pyjwt~=2.10",
    "pymongo~=4.13",
    "pymysql~=1.1",
    "pyyaml~=6.0",
    "sentence-transformers~=4.1",
    "sqlalchemy~=2.0",
    "sqlmodel~=0.0",
    "strawberry-graphql[asgi]~=0.275",
    "uvicorn~=0.34",
]

[dependency-groups]
dev = [
    "asyncpg~=0.30",
    "dagster-cloud~=1.11",
    "dagster-dg-cli~=1.11",
    "dagster-webserver~=1.11",
    "greenlet~=3.2",
    "mysql-connector-python~=8.3",
    "passlib[bcrypt]~=1.7",
    "polylith-cli~=1.30",
    "pre-commit~=4.2",
    "psycopg2-binary~=2.9",
    "pytest~=8.4",
    "pytest-asyncio~=1.1",
    "ruff~=0.11",
    "strawberry-graphql[debug-server]~=0.275",
    "testcontainers[postgres]~=4.13",
    "ty~=0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
dev-mode-dirs = ["components", "bases", "development", "."]

[tool.uv]
dev-dependencies = ["polylith-cli"]

[tool.polylith.bricks]
"bases/lif/query_cache_module" = "lif/query_cache_module"
"bases/lif/query_cache_restapi" = "lif/query_cache_restapi"
"bases/lif/query_planner_restapi" = "lif/query_planner_restapi"
"components/lif/datatypes" = "lif/datatypes"
"components/lif/mongodb_connection" = "lif/mongodb_connection"
"components/lif/query_cache_read" = "lif/query_cache_read"
"components/lif/query_cache_read_store_mongodb" = "lif/query_cache_read_store_mongodb"
"components/lif/query_cache_read_store_in_memory" = "lif/query_cache_read_store_in_memory"
"components/lif/query_cache_service" = "lif/query_cache_service"
"components/lif/query_planner_service" = "lif/query_planner_service"
"components/lif/logging" = "lif/logging"
"bases/lif/api_graphql" = "lif/api_graphql"
"components/lif/openapi_to_graphql" = "lif/openapi_to_graphql"
"components/lif/string_utils" = "lif/string_utils"
"components/lif/mdr_client" = "lif/mdr_client"
"components/lif/mdr_services" = "lif/mdr_services"
"bases/lif/semantic_search_mcp_server" = "lif/semantic_search_mcp_server"
"components/lif/semantic_search_service" = "lif/semantic_search_service"
"components/lif/openapi_schema_parser" = "lif/openapi_schema_parser"
"bases/lif/advisor_restapi" = "lif/advisor_restapi"
"components/lif/langchain_agent" = "lif/langchain_agent"
"components/lif/auth" = "lif/auth"
"components/lif/advisor_service" = "lif/advisor_service"
"components/lif/exceptions" = "lif/exceptions"
"components/lif/composer" = "lif/composer"
"bases/lif/orchestrator_restapi" = "lif/orchestrator_restapi"
"components/lif/orchestrator_service" = "lif/orchestrator_service"
"components/lif/orchestrator_query_plan_parser" = "lif/orchestrator_query_plan_parser"
"components/lif/orchestrator_clients" = "lif/orchestrator_clients"
"bases/lif/identity_mapper_restapi" = "lif/identity_mapper_restapi"
"components/lif/identity_mapper_service" = "lif/identity_mapper_service"
"components/lif/identity_mapper_storage" = "lif/identity_mapper_storage"
"components/lif/identity_mapper_storage_sql" = "lif/identity_mapper_storage_sql"
"components/lif/data_source_adapters" = "lif/data_source_adapters"
"bases/lif/example_data_source_rest_api" = "lif/example_data_source_rest_api"
"bases/lif/mdr_restapi" = "lif/mdr_restapi"
"bases/lif/translator_restapi" = "lif/translator_restapi"
"components/lif/mdr_auth" = "lif/mdr_auth"
"components/lif/example_data_source_service" = "lif/example_data_source_service"
"components/lif/translator" = "lif/translator"
"components/lif/mdr_utils" = "lif/mdr_utils"
"components/lif/mdr_dto" = "lif/mdr_dto"
"components/lif/schema" = "lif/schema"
"components/lif/utils" = "lif/utils"
"components/lif/dynamic_models" = "lif/dynamic_models"
"components/lif/openapi_schema" = "lif/openapi_schema"
"components/lif/graphql" = "lif/graphql"

[tool.ruff]
line-length = 120
required-version = "~=0.11.13"

[tool.ruff.format]
docstring-code-format = true
# use '\n' line endings for all files (*nix/mac default)
line-ending = "lf"
skip-magic-trailing-comma = true

[tool.ruff.lint]
ignore = [
  "E711",
  "E712",
  "F811",
  "F841",  # Eventually should be removed and the code cleaned up
  "TD002",
  "TD003"
]

[tool.ty.src]
exclude = [
  "scripts/",
  "test/",
  "bases/lif/advisor_restapi/",
  "bases/lif/api_graphql/",
  "bases/lif/mdr_restapi/",
  "bases/lif/semantic_search_mcp_server",
  "components/lif/langchain_agent/",
  "components/lif/mdr_dto/",
  "components/lif/mdr_services/",
  "components/lif/mdr_utils/",
  "components/lif/openapi_to_graphql/",
  "components/lif/orchestrator_clients/",
  "components/lif/semantic_search_service/",
  "development/"
]

